requirements.txt
numpy==1.22.1
pandas==1.3.5
python-binance==1.0.15